// packages/api/ku-key/models/today-course-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * KU-KEY API
 * API for KU-KEY service
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface TodayCourseDto
 */
export interface TodayCourseDto {
  /**
   * 강의 이름
   * @type {string}
   * @memberof TodayCourseDto
   */
  courseName: string
  /**
   * 교수님 성함
   * @type {string}
   * @memberof TodayCourseDto
   */
  professorName: string
  /**
   * 강의실
   * @type {string}
   * @memberof TodayCourseDto
   */
  classroom?: string
  /**
   * 시작 시간
   * @type {string}
   * @memberof TodayCourseDto
   */
  startTime?: string
  /**
   * 종료 시간
   * @type {string}
   * @memberof TodayCourseDto
   */
  endTime?: string
}